8346386: [CRaC] Move CRaC’s Java code out of restricted packages
