+++ src/hotspot/share/classfile/vmSymbols.hpp
@@ -744,7 +744,7 @@ class SerializeClosure;
   template(toFileURL_signature,                             "(Ljava/lang/String;)Ljava/net/URL;")                 \
   template(url_array_classloader_void_signature,            "([Ljava/net/URL;Ljava/lang/ClassLoader;)V")          \
                                                                                                                   \
-  template(jdk_crac_Core,                          "jdk/crac/Core")                                               \
+  template(jdk_crac_Core,                          "jdk/internal/crac/mirror/Core")                               \
   template(checkpointRestoreInternal_name,         "checkpointRestoreInternal")                                   \
   template(checkpointRestoreInternal_signature,    "(J)Ljava/lang/String;")                                       \
                                                                                                                   \
