+++ src/java.base/linux/classes/sun/nio/ch/EPollSelectorImpl.java
@@ -26,8 +26,8 @@
 
 package sun.nio.ch;
 
-import jdk.crac.Context;
-import jdk.crac.Resource;
+import jdk.internal.crac.mirror.Context;
+import jdk.internal.crac.mirror.Resource;
 import jdk.internal.access.JavaIOFileDescriptorAccess;
 import jdk.internal.access.SharedSecrets;
 import jdk.internal.crac.Core;
