+++ src/java.base/share/classes/java/io/FileOutputStream.java
@@ -29,8 +29,8 @@ import java.nio.channels.FileChannel;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
-import jdk.crac.Context;
-import jdk.crac.Resource;
+import jdk.internal.crac.mirror.Context;
+import jdk.internal.crac.mirror.Resource;
 import jdk.internal.crac.Core;
 import jdk.internal.crac.JDKResource;
 import jdk.internal.crac.OpenResourcePolicies;
