+++ src/java.base/share/classes/java/lang/Process.java
@@ -25,9 +25,9 @@
 
 package java.lang;
 
-import jdk.crac.Context;
-import jdk.crac.Resource;
-import jdk.crac.impl.CheckpointOpenResourceException;
+import jdk.internal.crac.mirror.Context;
+import jdk.internal.crac.mirror.Resource;
+import jdk.internal.crac.mirror.impl.CheckpointOpenResourceException;
 import jdk.internal.crac.Core;
 import jdk.internal.crac.JDKFdResource;
 import jdk.internal.crac.OpenResourcePolicies;
