+++ src/java.base/share/classes/java/lang/ProcessBuilder.java
@@ -33,8 +33,8 @@ import java.io.InputStream;
 import java.io.OutputStream;
 import java.util.*;
 
-import jdk.crac.Context;
-import jdk.crac.Resource;
+import jdk.internal.crac.mirror.Context;
+import jdk.internal.crac.mirror.Resource;
 import jdk.internal.access.JavaIOFileDescriptorAccess;
 import jdk.internal.access.SharedSecrets;
 import jdk.internal.crac.Core;
