+++ src/java.base/share/classes/java/lang/ref/Reference.java
@@ -25,8 +25,8 @@
 
 package java.lang.ref;
 
-import jdk.crac.Context;
-import jdk.crac.Resource;
+import jdk.internal.crac.mirror.Context;
+import jdk.internal.crac.mirror.Resource;
 import jdk.internal.crac.Core;
 import jdk.internal.crac.JDKResource;
 import jdk.internal.misc.Unsafe;
