+++ src/java.base/share/classes/jdk/internal/crac/JDKFdResource.java
@@ -1,7 +1,7 @@
 package jdk.internal.crac;
 
-import jdk.crac.Context;
-import jdk.crac.Resource;
+import jdk.internal.crac.mirror.Context;
+import jdk.internal.crac.mirror.Resource;
 import sun.security.action.GetBooleanAction;
 
 import java.util.function.Supplier;
