+++ src/java.base/share/classes/jdk/internal/crac/mirror/impl/CheckpointOpenResourceException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, 2021, Azul Systems, Inc. All rights reserved.
+ * Copyright (c) 2017, 2021, Azul Systems, Inc. All rights reserved.
  * Copyright (c) 2021, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
@@ -24,37 +24,12 @@
  * questions.
  */
 
-package javax.crac;
+package jdk.internal.crac.mirror.impl;
 
-/**
- * A {@code Resource} that allows other {@code Resource}s to be registered with it.
- *
- * <p>{@code Context} implementation overrides {@code beforeCheckpoint} and {@code afterRestore}, defining how the notification about checkpoint and restore will be distributed by the {@code Context} hierarchy.
- *
- * <p>A {@code Context} implementor is encouraged to respect properties of the global {@code Context}.
- *
- * @param <R> Resource
- */
-public abstract class Context<R extends Resource> implements Resource {
+public class CheckpointOpenResourceException extends Exception {
+    private static final long serialVersionUID = -3858375642480846931L;
 
-    /** Creates a {@code Context}.
-     */
-    protected Context() {
+    public CheckpointOpenResourceException(String details, Throwable cause) {
+        super(details, cause);
     }
-
-    @Override
-    public abstract void beforeCheckpoint(Context<? extends Resource> context)
-            throws CheckpointException;
-
-    @Override
-    public abstract void afterRestore(Context<? extends Resource> context)
-            throws RestoreException;
-
-    /**
-     * Registers a {@code Resource} with this {@code Context}.
-     *
-     * @param resource {@code Resource} to be registered.
-     * @throws NullPointerException if {@code resource} is {@code null}
-     */
-    public abstract void register(R resource);
 }
