+++ src/java.base/share/classes/jdk/internal/crac/mirror/impl/ExceptionHolder.java
@@ -1,4 +1,4 @@
-package jdk.crac.impl;
+package jdk.internal.crac.mirror.impl;
 
 import java.util.function.Supplier;
