+++ src/java.base/share/classes/jdk/internal/crac/mirror/impl/KeepAlive.java
@@ -20,7 +20,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package jdk.crac.impl;
+package jdk.internal.crac.mirror.impl;
 
 import java.util.concurrent.CountDownLatch;
