+++ src/java.base/share/classes/jdk/internal/crac/mirror/impl/OrderedContext.java
@@ -24,9 +24,12 @@
  * questions.
  */
 
-package jdk.crac.impl;
+package jdk.internal.crac.mirror.impl;
 
-import jdk.crac.*;
+import jdk.internal.crac.mirror.Context;
+import jdk.internal.crac.mirror.Resource;
+import jdk.internal.crac.mirror.RestoreException;
+import jdk.internal.crac.mirror.impl.AbstractContext;
 
 import java.util.*;
