+++ src/java.base/share/classes/jdk/internal/crac/mirror/package-info.java
@@ -94,5 +94,4 @@ restoring instances from the image.
  * @since TBD
  */
 
-package javax.crac;
-
+package jdk.internal.crac.mirror;
