+++ src/java.base/share/classes/jdk/internal/util/jar/PersistentJarFile.java
@@ -25,8 +25,8 @@
 
 package jdk.internal.util.jar;
 
-import jdk.crac.Context;
-import jdk.crac.Resource;
+import jdk.internal.crac.mirror.Context;
+import jdk.internal.crac.mirror.Resource;
 import jdk.internal.access.SharedSecrets;
 import jdk.internal.crac.Core;
 import jdk.internal.crac.LoggerContainer;
