+++ src/java.base/share/classes/sun/net/www/protocol/jar/JarFileFactory.java
@@ -38,8 +38,8 @@ import jdk.internal.util.OperatingSystem;
 import sun.net.util.URLUtil;
 
 import jdk.internal.crac.Core;
-import jdk.crac.Context;
-import jdk.crac.Resource;
+import jdk.internal.crac.mirror.Context;
+import jdk.internal.crac.mirror.Resource;
 
 /* A factory for cached JAR file. This class is used to both retrieve
  * and cache Jar files.
