+++ src/java.base/share/classes/sun/nio/ch/AsynchronousFileChannelImpl.java
@@ -25,8 +25,8 @@
 
 package sun.nio.ch;
 
-import jdk.crac.Context;
-import jdk.crac.Resource;
+import jdk.internal.crac.mirror.Context;
+import jdk.internal.crac.mirror.Resource;
 import jdk.internal.crac.OpenResourcePolicies;
 import jdk.internal.crac.JDKFileResource;
