+++ src/java.base/share/classes/sun/security/provider/FileInputStreamPool.java
@@ -25,8 +25,8 @@
 
 package sun.security.provider;
 
-import jdk.crac.Context;
-import jdk.crac.Resource;
+import jdk.internal.crac.mirror.Context;
+import jdk.internal.crac.mirror.Resource;
 import jdk.internal.crac.Core;
 import jdk.internal.crac.JDKResource;
