+++ src/jdk.crac/share/classes/jdk/crac/Core.java
@@ -1,0 +1,88 @@
+/*
+ * Copyright (c) 2017, 2021, Azul Systems, Inc. All rights reserved.
+ * Copyright (c) 2021, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.  Oracle designates this
+ * particular file as subject to the "Classpath" exception as provided
+ * by Oracle in the LICENSE file that accompanied this code.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ */
+
+package jdk.crac;
+
+import jdk.internal.crac.mirror.impl.BlockingOrderedContext;
+import jdk.internal.crac.mirror.impl.OrderedContext;
+
+/**
+ * The coordination service.
+ */
+public class Core {
+
+    /** This class is not instantiable. */
+    private Core() {
+    }
+
+    private static final Context<Resource> globalContext = new ContextWrapper(
+        jdk.internal.crac.mirror.impl.GlobalContext.createGlobalContextImpl());
+
+    static {
+        jdk.internal.crac.mirror.Core.getGlobalContext().register(new ResourceWrapper(null, globalContext));
+    }
+
+    /**
+     * Gets the global {@code Context} for checkpoint/restore notifications.
+     *
+     * @return the global {@code Context}
+     */
+    public static Context<Resource> getGlobalContext() {
+        return globalContext;
+    }
+
+    /**
+     * Requests checkpoint and returns upon a successful restore.
+     * May throw an exception if the checkpoint or restore are unsuccessful.
+     *
+     * @throws CheckpointException if an exception occured during checkpoint
+     * notification and the execution continues in the original Java instance.
+     * @throws RestoreException if an exception occured during restore
+     * notification and execution continues in a new Java instance.
+     * @throws UnsupportedOperationException if checkpoint/restore is not
+     * supported, no notification performed and the execution continues in
+     * the original Java instance.
+     */
+    public static void checkpointRestore() throws
+            CheckpointException,
+            RestoreException {
+        try {
+            jdk.internal.crac.mirror.Core.checkpointRestore();
+        } catch (jdk.internal.crac.mirror.CheckpointException e) {
+            CheckpointException newException = new CheckpointException();
+            for (Throwable t : e.getSuppressed()) {
+                newException.addSuppressed(t);
+            }
+            throw newException;
+        } catch (jdk.internal.crac.mirror.RestoreException e) {
+            RestoreException newException = new RestoreException();
+            for (Throwable t : e.getSuppressed()) {
+                newException.addSuppressed(t);
+            }
+            throw newException;
+        }
+    }
+}
