+++ src/jdk.crac/share/classes/jdk/crac/package-info.java
@@ -95,4 +95,3 @@ restoring instances from the image.
  */
 
 package jdk.crac;
-
